\name{DeducerExtras}
\alias{Param}
\alias{ParamAny}
\alias{ParamCharacter}
\alias{ParamLogical}
\alias{ParamMultipleVariables}
\alias{ParamNumeric}
\alias{ParamRFunctionResult}
\alias{ParamRObject}
\alias{ParamVariable}
\alias{RFunction}
\alias{RFunctionDialog}
\alias{RFunctionList}
\alias{RFunctionListDialog}
\title{
  Imports from org.rosuda.deducer.widgets.param
}
\description{
Java imports
}